sprintF now uses double

+++ src/lib/convert/Convert.cpp
@@ -4,7 +4,7 @@
 #include "../../Common.h"
 #include "Convert.h"
 
-void sprintF(char *result, const char *source, float f) {
+void sprintF(char *result, const char *source, double f) {
   bool ok = false;
   uint8_t mas = 0, frac = 0, len = 0;
   char* a;
